ft_printf project almost done, needs minor fixes

+++ ft_printf.h
@@ -6,15 +6,16 @@
 /*   By: dvlachos <dvlachos@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/25 15:26:12 by dvlachos          #+#    #+#             */
-/*   Updated: 2024/11/27 17:32:50 by dvlachos         ###   ########.fr       */
+/*   Updated: 2024/11/29 16:37:41 by dvlachos         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef FT_PRINTF_H
 # define FT_PRINTF_H
 # include <unistd.h>
-# include <stdarg.h>
+# include <stdint.h>
 # include <stdio.h>
+# include <limits.h>
 
 int	ft_printf(const char *c, ...);
 int	prstr(char *str);
@@ -22,7 +23,7 @@ int	prpointer(void *poi);
 int	prnum(int n);
 int	prhex(unsigned long num, char x);
 int	prchar(int c);
-int	ft_strlen(const char *str);
+size_t	ft_strlen(const char *str);
 int	prunsint(unsigned int n);
 
 #endif
